+++ src/feature/restaurant/restaurant.module.ts
@@ -3,12 +3,11 @@ import { RestaurantService } from './restaurant.service';
 import { RestaurantController } from './restaurant.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { Restaurant } from 'src/entity/restaurant.entity';
-import { ReviewService } from './review.service';
-import { Review } from 'src/entity/review.entity';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Restaurant, Review])],
+  imports: [TypeOrmModule.forFeature([Restaurant])],
   controllers: [RestaurantController],
-  providers: [RestaurantService, ReviewService],
+  providers: [RestaurantService],
+  exports: [RestaurantService],
 })
 export class RestaurantModule {}
